refactor : 최상위 패키지 laddergame 생성
