+++ src/main/java/laddergame/util/TrueOrFalseGenerator.java
@@ -1,4 +1,4 @@
-package util;
+package laddergame.util;
 
 public interface TrueOrFalseGenerator {
     boolean generate();
