+++ src/test/java/laddergame/domain/HeightTest.java
@@ -1,6 +1,6 @@
-package domain;
+package laddergame.domain;
 
-import domain.Height;
+import laddergame.Height;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
